feature: Remove template
