+++ Sources/Subs-MessageIndex.php
@@ -88,4 +88,27 @@ function getBoardList($boardListOptions = array())
 	return $return_value;
 }
 
+
+function child_boards($board) {
+	// Sort the links into an array with new boards bold so it can be imploded.
+	$children = array();
+	/* Each child in each board's children has:
+			id, name, description, new (is it new?), topics (#), posts (#), href, link, and last_post. */
+	foreach ($board['children'] as $child)
+	{
+		if (!$child['is_redirect'])
+			$child['link'] = '<a href="' . $child['href'] . '" ' . ($child['new'] ? 'class="board_new_posts" ' : '') . 'title="' . ($child['new'] ? $txt['new_posts'] : $txt['old_posts']) . ' (' . $txt['board_topics'] . ': ' . comma_format($child['topics']) . ', ' . $txt['posts'] . ': ' . comma_format($child['posts']) . ')">' . $child['name'] . ($child['new'] ? '</a> <a  ' . ($child['new'] ? 'class="new_posts" ' : '') . 'href="' . $scripturl . '?action=unread;board=' . $child['id'] . '" title="' . $txt['new_posts'] . ' (' . $txt['board_topics'] . ': ' . comma_format($child['topics']) . ', ' . $txt['posts'] . ': ' . comma_format($child['posts']) . ')"><span class="new_posts">' . $txt['new'] . '</span>' : '') . '</a>';
+		else
+			$child['link'] = '<a href="' . $child['href'] . '" title="' . comma_format($child['posts']) . ' ' . $txt['redirects'] . '">' . $child['name'] . '</a>';
+
+		// Has it posts awaiting approval?
+		if ($child['can_approve_posts'] && ($child['unapproved_posts'] | $child['unapproved_topics']))
+			$child['link'] .= ' <a href="' . $scripturl . '?action=moderate;area=postmod;sa=' . ($child['unapproved_topics'] > 0 ? 'topics' : 'posts') . ';brd=' . $child['id'] . ';' . $context['session_var'] . '=' . $context['session_id'] . '" title="' . sprintf($txt['unapproved_posts'], $child['unapproved_topics'], $child['unapproved_posts']) . '" class="moderation_link">(!)</a>';
+
+		$children[] = $child['new'] ? '<strong>' . $child['link'] . '</strong>' : $child['link'];
+	}
+	
+	return implode(',', $children);
+}
+
 ?>
